code refactoring , added aggrid separate component
