+++ railpass/src/components/mainPages/timing/timing.tsx
@@ -1,27 +1,7 @@
-import { AgGridReact } from "ag-grid-react";
 import React, { useState } from "react";
 import { useParams, useLocation } from "react-router-dom";
-
-const extractTrainData = (trainData: any) => {
-  const extractedTrainData = trainData.data.map(
-    (train: {
-      from_sta: any;
-      to_sta: any;
-      train_number: any;
-      train_name: any;
-      from_station_name: any;
-      to_station_name: any;
-    }) => ({
-      train_number: train.train_number,
-      train_name: train.train_name,
-      from_station: train.from_station_name,
-      to_station: train.to_station_name,
-      departure_time: train.from_sta,
-      arrival_time: train.to_sta,
-    })
-  );
-  return extractedTrainData;
-};
+import { extractTrainData } from "../../../services/utils";
+import { Grid } from "../../reusableComponents/agGrid";
 
 const Timing = () => {
   // const { stationName } = useParams();
@@ -79,7 +59,7 @@ const Timing = () => {
       className="ag-theme-quartz"
       style={{ height: "100%", width: "100%", textAlign: "center" }}
     >
-      <AgGridReact rowData={rowData} pagination={false} columnDefs={colDefs} />
+      <Grid rowData={rowData} pagination={false} columnDefs={colDefs} />
     </div>
   );
 };
