+++ railpass/src/components/reusableComponents/agGrid/index.tsx
@@ -1,0 +1,2 @@
+import Grid from "./agGrid";
+export {Grid};
